forward declaration to fix safety profile builds
